newsfeed/requirements.txt
aylien-apiclient==0.7.0
certifi==2019.6.16
chardet==3.0.4
Django==1.11
django-admin-view-permission==1.9
django-bootstrap3==11.1.0
django-discover-runner==1.0
django-rest-framework==0.1.0
django-role-permissions==2.2.1
django-tables2==2.1.0
djangorestframework==3.10.2
httplib2==0.13.1
idna==2.8
psycopg2==2.8.3
pytz==2019.2
requests==2.22.0
six==1.12.0
urllib3==1.25.3
